requirements.txt
streamlit>=1.33.0
streamlit-webrtc>=0.47.7
av==10.0.0
numpy>=1.24.0
pydub>=0.25.1
imageio-ffmpeg>=0.4.9
soundfile>=0.12.1
openai-whisper>=20231117
torch>=2.1.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
altair>=5.3.0
pandas>=2.0.0
